Add symbol size to the SYMBOL response
The client may use this value to validate the answer if the requested
address is not at the beginning of the symbol, e.g.:

C: a013 SYMBOL c07070a2
S: * SYMBOL c07070a0 c0 R "linux_banner"
C: a013 OK SYMBOL completed

As you can see, the offset is 2, but without the size it would be
impossible to tell if the server actually answered our query (or a
different one).

Note that it does not refer to the actual size of the symbol, as found
in the ELF symbol table, but rather to the address range for which
the SYMBOL command returns this symbol.