Terminate all sessions on server shutdown
The logic revolves around term_state. Under normal conditions, the state
is conn_running, and the FSM operates as usual.

If the field is set to conn_bye_pending, the FSM will send out a BYE
message when it sends the next response. The previous message is
preserved (so you get a tagged response to the TERMINATE command) and
sent after the BYE message. The read part of the socket is shut down.

After that, the state is bumped to conn_bye_sent to prevent further BYE
messages (e.g. because of TERMINATE commands sent to other connections).

Note that do_TERMINATE doesn't call disconnect() directly, but since
all the state is "broadcast" to all connections, the originating
connection gets it, too.