Hook the BYE code in the main loop instead of conn_respond
While the hook in conn_respond() works fine for any busy connection,
it fails to send the message on an idle one (i.e. when it is waiting
for a command).

However, since we can't insert the BYE message anywhere in the stream
(e.g. it would be ignored/misunderstood inside a literal), we must wait
for a quiescent period.

The idle case can be "fixed" by checking in the main server loop (just
before polling), but this place doesn't work for busy connections,
because they will always have something to send. That's why we also
call it from handle_conn, i.e. right after raw I/O is finished and
before new output can be queued by a handler.

Note that we must preserve the wait events, because the code is now
hooked at a very low level.