guiserver: issue a message when the server starts
The idea is that you can start crash with stdandard input redirected
from a pipe. You can then issue the 'guiserver' command but the
connection cannot be open until the server is fully initialised.

To close the race condition, you can wait for the confirmation message
on the output pipe.