Clean up the use of getline_status
First, the gls field of struct conn is a write-only variable. The value
should have been used in conn_readcommand() to determine whether another
line can be read immediately, but after the API cleanup, this should be
done by calling getline_hasmore() instead.

As a result, the return value of fdgetline() is only relevant inside
conn_readcommand(), so let's declare it as a local variable there.